Support a placeholder file within the modules directory
This lets us check the modules directory into version control while
empty.

In general, the modules installed within the modules directory should be
ignored, but if you specify a specific module_path within the
Vagrantfile it must exist for vagrant to run. Setting up a placeholder
file within the modules directory lets us check it in and .gitignore the
rest of the directories installed within it.